+++ internal/chain/errors.go
@@ -5,3 +5,4 @@ import (
 )
 
 var ErrChainRequestUnreachable = errors.New("chain request unreachable")
+var ErrEstimateFee = errors.New("cannot estimate fee")